In Python 2.3, range() cares that it was passed integer types, so pass an integer to avoid an unhelpful deprecation warning.  (Before Python 2.3, range() and many other built-in functions accepted floats as ints "by accident"; those cases now generate warnings when passes a float.)

+++ lib/python/ZServer/medusa/test/max_sockets.py
@@ -50,7 +50,7 @@ def max_select_sockets():
     while 1:
         try:
             num = len(sl)
-            for i in range(1 + len(sl) * 0.05):
+            for i in range(1 + int(len(sl) * 0.05)):
                     # Increase exponentially.
                 s = socket.socket (socket.AF_INET, socket.SOCK_STREAM)
                 s.bind (('',0))
